+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## change log ##
 
 * 2017-02-14:  3.2.1
+  - m2r.saveM2RModel: remove duplicates in each mapping group
   - encode `#` -> `%23` in "Resolve with" email upon ontology registration
   
   - \#35: "API operation to insert new terms in vocabulary"